refactor: move transaction module to kernel
