+++ crates/core/src/kernel/mod.rs
@@ -10,6 +10,7 @@ pub mod error;
 pub mod models;
 pub mod scalars;
 mod snapshot;
+pub mod transaction;
 
 pub use error::*;
 pub use models::*;
