+++ crates/core/src/kernel/transaction/conflict_checker.rs
