+++ crates/core/src/kernel/transaction/state.rs
