+++ crates/core/src/operations/set_tbl_properties.rs
@@ -5,8 +5,8 @@ use std::sync::Arc;
 
 use futures::future::BoxFuture;
 
-use super::transaction::{CommitBuilder, CommitProperties};
 use super::{CustomExecuteHandler, Operation};
+use crate::kernel::transaction::{CommitBuilder, CommitProperties};
 use crate::kernel::Action;
 use crate::logstore::LogStoreRef;
 use crate::protocol::DeltaOperation;
