+++ crates/core/tests/command_merge.rs
@@ -7,9 +7,9 @@ use datafusion::dataframe::DataFrame;
 use datafusion::prelude::SessionContext;
 use datafusion_common::Column;
 use datafusion_expr::{col, lit, Expr};
+use deltalake_core::kernel::transaction::TransactionError;
 use deltalake_core::kernel::{DataType as DeltaDataType, PrimitiveType, StructField, StructType};
 use deltalake_core::operations::merge::MergeMetrics;
-use deltalake_core::operations::transaction::TransactionError;
 use deltalake_core::protocol::SaveMode;
 use deltalake_core::{open_table, DeltaOps, DeltaResult, DeltaTable, DeltaTableError};
 use std::sync::Arc;
